fixed selected subtitle
