+++ jukebox-app/src/main/java/se/qxx/android/jukebox/activities/SubSelectActivity.java
@@ -17,14 +17,12 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.AdapterView.OnItemClickListener;
 import android.widget.ListView;
 
 import com.google.android.gms.cast.framework.CastContext;
 import com.google.android.gms.cast.framework.media.RemoteMediaClient;
 
-public class SubSelectActivity extends AppCompatActivity implements OnItemClickListener, OnDismissListener {
+public class SubSelectActivity extends AppCompatActivity implements OnDismissListener {
     private CastContext mCastContext;
 
     @Override
@@ -61,47 +59,18 @@ public class SubSelectActivity extends AppCompatActivity implements OnItemClickL
                 JukeboxSettings.get().getServerPort());
 
         final ListView v = (ListView)findViewById(R.id.listSubtitlePicker);
-		v.setOnItemClickListener(this);
 
         jh.listSubtitles(
             this.getMedia(),
 				(response) -> {
             		runOnUiThread(() -> {
-						SubtitleLayoutAdapter adapter = new SubtitleLayoutAdapter(this, response.getSubtitleList());
-						v.setAdapter(adapter);
+						SubtitleLayoutAdapter adapter = new SubtitleLayoutAdapter(this, getMedia().getID(), response.getSubtitleList());
+                        v.setOnItemClickListener(adapter);
+                        v.setAdapter(adapter);
 					});
 				});
 
     }
-
-    @Override
-	public void onItemClick(AdapterView<?> arg0, View arg1, int arg2, long arg3) {
-		final Subtitle sub = (Subtitle)arg0.getItemAtPosition(arg2);	
-		Logger.Log().d(String.format("Setting subtitle to %s", sub.getDescription()));
-
-		if (ChromeCastConfiguration.isChromeCastActive()) {
-			RemoteMediaClient client = ChromeCastConfiguration.getRemoteMediaClient(this.getApplicationContext());
-
-            if (client != null) {
-            	client.setActiveMediaTracks(new long[] {(long)arg2});
-            }
-        }
-		else {
-			final JukeboxConnectionHandler jh = new JukeboxConnectionHandler(
-					JukeboxSettings.get().getServerIpAddress(),
-					JukeboxSettings.get().getServerPort(),
-					JukeboxConnectionProgressDialog.build(this, "Setting subtitle ..."));
-
-			Thread t = new Thread(() ->
-                    jh.setSubtitle(
-                            JukeboxSettings.get().getCurrentMediaPlayer(),
-                            getMedia(),
-                            sub));
-			t.run();
-
-		}
-
-	}
 
 	@Override
 	public void onDismiss(DialogInterface dialog) {
