+++ jukebox-app/src/main/java/se/qxx/android/jukebox/adapters/support/SubtitleLayoutAdapter.java
@@ -2,8 +2,12 @@ package se.qxx.android.jukebox.adapters.support;
 
 import se.qxx.android.jukebox.activities.IncludeSubtitleRating;
 import se.qxx.android.jukebox.R;
+import se.qxx.android.jukebox.cast.ChromeCastConfiguration;
+import se.qxx.android.jukebox.dialogs.JukeboxConnectionProgressDialog;
 import se.qxx.android.jukebox.model.Model;
+import se.qxx.android.jukebox.settings.JukeboxSettings;
 import se.qxx.android.tools.Logger;
+import se.qxx.jukebox.comm.client.JukeboxConnectionHandler;
 import se.qxx.jukebox.domain.JukeboxDomain.Subtitle;
 import se.qxx.jukebox.domain.Sorter;
 
@@ -11,19 +15,28 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.AdapterView;
 import android.widget.BaseAdapter;
+
+import com.google.android.gms.cast.framework.media.RemoteMediaClient;
 
 import java.util.List;
 
-public class SubtitleLayoutAdapter extends BaseAdapter {
+public class SubtitleLayoutAdapter extends BaseAdapter implements AdapterView.OnItemClickListener {
 
 	private Context context;
+	private int mediaId;
 	private List<Subtitle> sortedSubtitles = null;
+	private int selectedSubId = 0;
 
-	public SubtitleLayoutAdapter(Context context, List<Subtitle> subtitles) {
+	public SubtitleLayoutAdapter(Context context, int mediaId, List<Subtitle> subtitles) {
 		super();
 		this.context = context;
+		this.mediaId = mediaId;
 		this.sortedSubtitles = Sorter.sortSubtitlesByRating(subtitles);
+
+		if (this.sortedSubtitles.size() > 0)
+		    this.selectedSubId = this.sortedSubtitles.get(0).getID();
 	}
 	
 	@Override
@@ -36,7 +49,10 @@ public class SubtitleLayoutAdapter extends BaseAdapter {
 	            v = vi.inflate(R.layout.subtitleitem, null);
 	        }
 	        Subtitle sub = (Subtitle)this.getItem(position);
-	        
+
+	        if (sub.getID() != this.selectedSubId)
+	            v.findViewById(R.id.imgSelected).setVisibility(View.GONE);
+
 	        IncludeSubtitleRating.initialize(sub, v);
 		}
 		catch (Exception e) {
@@ -61,5 +77,32 @@ public class SubtitleLayoutAdapter extends BaseAdapter {
 		return this.sortedSubtitles.get(position).hashCode();
 	}
 
+	@Override
+	public void onItemClick(AdapterView<?> arg0, View arg1, int arg2, long arg3) {
+		final Subtitle sub = (Subtitle) arg0.getItemAtPosition(arg2);
+
+		Logger.Log().d(String.format("Setting subtitle to %s", sub.getDescription()));
+
+
+		if (ChromeCastConfiguration.isChromeCastActive()) {
+			RemoteMediaClient client = ChromeCastConfiguration.getRemoteMediaClient(this.context);
+
+			if (client != null) {
+				client.setActiveMediaTracks(new long[]{(long) arg2});
+			}
+		} else {
+			final JukeboxConnectionHandler jh = new JukeboxConnectionHandler(
+					JukeboxSettings.get().getServerIpAddress(),
+					JukeboxSettings.get().getServerPort());
+
+			Thread t = new Thread(() ->
+					jh.setSubtitle(
+							JukeboxSettings.get().getCurrentMediaPlayer(),
+							this.mediaId,
+							sub));
+			t.run();
+
+		}
+	}
+
 }
-
